Move Singleton to utils
